Initial commit - Add book collection app source code
This commit adds the initial source code for the book collection app. It includes the Book and BookCollection classes, as well as the data storage mechanism using a JSON file. The app can be run from the command line, and includes the ability to add a book to the collection, remove a book from the collection, and display the entire collection.

The source code is written in Python 3 and uses the datetime, os, json, and collections modules.